Arreglo del monto total de la venta y relaciones

+++ database/migrations/2024_04_06_230254_create_sales_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('order_number')->unique();
             $table->string('payment_method');
-            $table->float('total');
+            $table->decimal('total', 10, 2);
             $table->softDeletes();
             $table->timestamps();
         });
